Move tests into separate file
